factored out find_min functions and iterators
